% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_infer.R
\name{CNV_infer}
\alias{CNV_infer}
\title{CNV_infer function}
\usage{
CNV_infer(
  ss.expr,
  mean.ctrl,
  gen.Loc,
  resolution,
  baseline_adj = FALSE,
  baseline = 0,
  chr.n,
  P12,
  mat.fab
)
}
\arguments{
\item{ss.expr}{scRNA-seq based expression for each test/control cell}

\item{mean.ctrl}{The average gene expression of control cells}

\item{resolution}{Adjusts the resolution, nrow(MSC)/(50*sharpness), used for the sciCNV-curve calculation. Default sharpness is =1.0 (best sharpnesses range between 0.6-1.4).}

\item{baseline_adj}{The baseline adjustment is only applied to test cells if it is TRUE. Default is FALSE.}

\item{baseline}{An optional correction to adjust the CNV zero setpoint (copy number gain =0) which is otherwise the median CNV of all genes.}

\item{chr.n}{List of chromosome numbers associated with the list of genes}

\item{P12}{The variable which is associated to resolution, is automatically calculated and is used for CNV calling.}

\item{mat.fab}{is a merged matrix of FF, AW, BD vectors per cell. These vectors are used to calculate the relative expression of test to control.}
}
\value{
The output is the sciCNV curve of each single-cell across entire genome
}
\description{
To infer copy number variation (sciCNV) at single-cell resolution
}
\note{
Please see the reference and supplmental materials described in the README file for additional information.
}
\examples{
iCNV_percell <- CNV_infer(ss.expr=norm_expr_percell, mean.ctrl, gen.Loc, resolution=50, chr.n, P12, mat.fab)

}
\author{
Ali Mahdipour-Shirayeh, Princess Margaret Cancer centre, University of Toronto
}
