% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_score.R
\name{CNV_score}
\alias{CNV_score}
\title{CNV_score function}
\usage{
CNV_score(M_nf)
}
\arguments{
\item{M_NF}{is a matrix of sciCNV profiles, with an extra column that reflects the average sciCNV profile of all suspected tumor cells (test cells)}
}
\value{
The tumor score (tumor likeness score) of each single-cell
}
\description{
Calculates a score for each cell reflecting the similarity of its sciCNV profile to the mean CNV profile of all suspected tumor cells
}
\note{
Please refer to the reference and supplemental materials described in the README for additional details.
}
