% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_break_glist.R
\name{heatmap_break_glist}
\alias{heatmap_break_glist}
\title{heatmap_break_glist function}
\usage{
heatmap_break_glist(CNV.mat2)
}
\arguments{
\item{CNV.mat2}{the sciCNV matrix of test and controcl cells}
}
\value{
The chromosome separation locations for the CNV heatmap against list of genes (CNV_heatmap_glist)
}
\description{
Finding separating lines among diverse chromosomes to be used in sketching heatmap
}
