% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_break_gloc.R
\name{heatmap_break_gloc}
\alias{heatmap_break_gloc}
\title{heatmap_break_gloc function}
\usage{
heatmap_break_gloc()
}
\value{
The chromosome separation locations for the CNV heatmap against genomic location (CNV_heatmap_gloc)
}
\description{
Finding separating lines among diverse chromosomes to be used in sketching heatmap
}
