% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mito_umi_gn.R
\name{Mito_umi_gn}
\alias{Mito_umi_gn}
\title{Mito_umi_gn function}
\usage{
Mito_umi_gn(mat, percent.mito.G, nUMI, nGene, No.test, drop.mads = 3)
}
\arguments{
\item{mat}{raw data matrix}

\item{percent.mito.G}{percentage of mitochondrial transcripts relative to total number of reads per cell}

\item{nUMI}{number of unified molecular identification (UMI)}

\item{nGene}{number of expressed genes}

\item{No.test}{number of test cells included in the data; can be used to delineate the populations of test and control cells in the heatmap}

\item{drop.mads}{number of the median absolute deviation (MAD) to drop damaged cells using the percentage of mitochondrial expression level}
}
\value{
The output represents damaged cells due to the given  for mitochondrial percentage cutoff
}
\description{
Function to sketch mitochondrial transcript content, nUMI and nGene data at cellular level to detect damaged cells within the population
}
\note{
Please refer to the reference and supplemental materials described in the README for additional details.
}
\examples{
damaged_cells <- Mito_umi_gn(mat=raw_data, percent.mito.G, nUMI, nGene, No.test=100, drop.mads = 3)

}
\author{
Ali Mahdipour-Shirayeh, Princess Margaret Cancer Centre, University of Toronto
}
