% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opt_MeanSD_RTAM2.R
\name{Opt_MeanSD_RTAM2}
\alias{Opt_MeanSD_RTAM2}
\title{Optimizing Function for RTAM2}
\usage{
Opt_MeanSD_RTAM2(
  Normalized_log,
  Order_Matrix,
  G_mtrx,
  nGene,
  Min_nGene,
  gene_cutoff
)
}
\arguments{
\item{Normalized_log}{the matrix of Log2(.+1)-normalized data}

\item{Order_Matrix}{the matrix of original order of gene expressions per cell}

\item{G_mtrx}{TPM-normalized matrix of the data}

\item{nGene}{number of expressed genes}

\item{Min_nGene}{minimum number of expressed genes across entire population}

\item{gene_cutoff}{the cutoff to calculate the summation of top (nonzero) gene expressions}
}
\value{
Calculates the CV (sd/mean) for the given nGene, Min_nGene and gene_cutoff applying RTAM2
}
\description{
An auxiliary function to detect the optimal number of top-expressed genes for scaling a data-set
}
