% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTAM_normalization.R
\name{RTAM_normalization}
\alias{RTAM_normalization}
\title{RTAM Nomalization Function}
\usage{
RTAM_normalization(mat, method, Min_nGn, Optimizing)
}
\arguments{
\item{mat}{The raw data matrix has genes as rows and cells as columns}

\item{Genes}{are annotated with their genomic location (start and end locations)}

\item{Order_Matrix}{is the original order of genes in the raw data}

\item{MinNumGenes}{is a threshold used to exclude poor QC cells that have few detectable genes (default= minimum of 250 genes).}

\item{OptNumGenes}{is the number of genes that when employed for RTAM normalization will optimally minimize the variance across cells of gene-set average expression}
}
\value{
Normalized data at a higher accuracy for all transcription data levels
}
\description{
RTAM1/2 normalization function designed but not limited to normalize scRNA-seq data
}
