% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaling_CNV.R
\name{Scaling_CNV}
\alias{Scaling_CNV}
\title{Scaling_sciCNV function}
\usage{
Scaling_CNV(V7Alt, n.TestCells, scaling.factor)
}
\arguments{
\item{V7Alt}{preliminary CNV matrix egnerated by sciCNV function}

\item{n.TestCells}{number of test cells}

\item{scaling.factor}{scaling factor to re-scale the preliminary CNV estimate}
}
\value{
The scaled sciCNV data to be matched with a benchmark (e.g. WES result)
}
\description{
To scale preliminary sciCNV-curves derived from scRNA-seq data
}
\note{
Please refer to the reference and supplemental materials described in the README for additional details
}
