% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sketch_AveCNV.R
\name{Sketch_AveCNV}
\alias{Sketch_AveCNV}
\title{Sketch_AveCNV function}
\usage{
Sketch_AveCNV(Ave.mat)
}
\arguments{
\item{Ave.mat}{the average of sciCNV-curves of test cells}
}
\value{
Sketched the average sciCNV of all cells (for instance that of tumor karyotype)
}
\description{
Plots the average pre-sciCNV profile of test cells by genomic location
}
