% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_htmp_gloc.R
\name{CNV_htmp_gloc}
\alias{CNV_htmp_gloc}
\title{CNV_htmp_gloc function}
\usage{
CNV_htmp_gloc(
  CNVmat,
  clustering = FALSE,
  clustering.type = NA,
  sorting = TRUE,
  CNVscore = NULL,
  cluster.lines = NULL,
  breakGloc,
  No.test
)
}
\arguments{
\item{CNVmat}{copy number variation matrix}

\item{clustering}{TRUE/FALSE variable specifying whether to cluster cells based on their CNV similarities. Default is "FALSE"}

\item{clustering.type}{Variable specifying the clustering method to be used in generating the heatmap. Possible options are "pearson",
"euclidean", " spearman", ... "original" (retains the original cell order without unsupervised clustering).
Default is "pearson". Only enabled when clustering = "TRUE"}

\item{sorting}{TRUE/FALSE variable specifying whether to sort cells based on their tumor CNV score from the largest to smallest tumor scores. Default is FALSE.}

\item{CNVscore}{is the tumor CNV score matrix for all cells (possibly ranked within clusters). Only used when sorting.clusters = TRUE.}

\item{cluster.lines}{is a list of values which can be used to separate cell clusters within the population; only used if multiple clusters present}

\item{breakGloc}{is a set of values each defines a vertical line that separates chromosomes}

\item{No.test}{number of test cells included in the data; can be used to delineate the populations of test and control cells in the heatmap}
}
\value{
The output is the heatmap of sciCNV matrix for test and control cells against genomic location
}
\description{
Generates heatmap of sciCNV profiles, plotted by Genomic location (gloc)
}
\note{
Please refer to the reference and supplemental materials described in the README for additional details.
}
\examples{
CNVmat <- system.file("extdata", "Sample_CNV_matrix.txt", package="sciCNV")
breakGloc <- system.file("extdata", "Sample_breakGloc.txt", package = "sciCNV")
CNV_htmp_gloc(CNVmat, breakGloc=breakGloc, sorting = FALSE,  No.test=20)

}
\author{
Ali Mahdipour-Shirayeh, Princess Margaret Cancer Centre, University of Toronto
}
