% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_score.R
\name{CNV_score}
\alias{CNV_score}
\title{CNV_score function}
\usage{
CNV_score(M_nf)
}
\arguments{
\item{M_nF}{is a matrix of sciCNV profiles, with an extra column that reflects the average
sciCNV profile of all suspected tumor cells (test cells)}
}
\value{
Calculates the tumor likeness score for each single-cell which is the similarity of sciCNV curve of each cell to average sciCNV of tet cells
}
\description{
Calculates a score for each cell reflecting the similarity of its sciCNV
profile to the mean CNV profile of all suspected tumor cells
}
\note{
Please see the reference and supplmental materials described in the README file for additional information.
}
\examples{
tumor_likeness_score <- CNV_score(M_nf=sciCNV_mat)

}
\author{
Ali Mahdipour-Shirayeh, Princess Margaret Cancer centre, University of Toronto
}
