% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_break_glist.R
\name{heatmap_break_glist}
\alias{heatmap_break_glist}
\title{heatmap_break_glist function}
\usage{
heatmap_break_glist(CNV.mat2)
}
\arguments{
\item{CNV.mat2}{the sciCNV matrix of test and controcl cells}
}
\value{
Calculates the seperation spots among chromosomes to sketch the heatmap of sciCNV matrix using CNV_htmp_glist function
}
\description{
Finding separating lines among diverse chromosomes to be used in sketching heatmap
}
\examples{
breakpoints_heatmap <- heatmap_break_glist(CNV.mat2 = normalized_data)

}
\author{
Ali Mahdipour-Shirayeh, Princess Margaret Cancer centre, University of Toronto
}
