% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_break_gloc.R
\name{heatmap_break_gloc}
\alias{heatmap_break_gloc}
\title{heatmap_break_gloc function}
\usage{
heatmap_break_gloc(CNV.mat2)
}
\arguments{
\item{CNV.mat2}{copy number variation matrix}
}
\value{
Calculates the seperation spots among chromosomes to sketch the heatmap of sciCNV matrix using CNV_htmp_gloc function
}
\description{
Finding separating lines among diverse chromosomes to be used in sketching heatmap
}
\examples{
breakpoints_heatmap <- heatmap_break_gloc()

}
\author{
Ali Mahdipour-Shirayeh, Princess Margaret Cancer centre, University of Toronto
}
