% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaling_CNV.R
\name{Scaling_CNV}
\alias{Scaling_CNV}
\title{Scaling_sciCNV function}
\usage{
Scaling_CNV(V7Alt, n.TestCells, scaling.factor)
}
\arguments{
\item{V7Alt}{preliminary CNV matrix egnerated by sciCNV function}

\item{Vn.TestCells}{number of test cells}

\item{Vscaling.factor}{scaling factor to re-scale the preliminary CNV estimate}
}
\value{
Scales sciCNV results to be matched with benchmark (e.g. WES result of the same test sample)
}
\description{
To scale preliminary sciCNV-curves derived from scRNA-seq data
}
\note{
Please refer to the reference and supplemental materials described in the README for additional details.
}
\examples{
scaled_sciCNV <- Scaling_CNV(V7Alt=sciCNV_mat, n.TestCells=100, scaling.factor=0.5)

}
\author{
Ali Mahdipour-Shirayeh, Princess Margaret Cancer Centre, University of Toronto
}
