% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sketch_AveCNV.R
\name{Sketch_AveCNV}
\alias{Sketch_AveCNV}
\title{Sketch_AveCNV function}
\usage{
Sketch_AveCNV(Ave.mat, Assoc.Chr)
}
\arguments{
\item{Ave.mat}{the average of sciCNV-curves of test cells}
}
\value{
Sketches the average sciCNV curve of all test cells; showing all CN gains/losses across entire genome
}
\description{
Plots the average pre-sciCNV profile of test cells by genomic location
}
\examples{
sketch_mean_sciCNV <- Sketch_AveCNV(Ave.mat=mean_sciCNV_mat)

}
\author{
Ali Mahdipour-Shirayeh, Princess Margaret Cancer Centre, University of Toronto
}
